requirements.txt
Cython==0.29.37
docker==7.1.0
circuitbreaker==2.0.0
concurrencytest==0.1.2
emoji==2.14.1
flake8==4.0.1
gcsfs==2024.6.0
google-api-core==2.22.0
google-cloud-core==2.4.1
google-auth==2.34.0
google-cloud-storage==2.17.0
google-cloud-aiplatform==1.108.0
google-cloud-bigquery==3.27.0
google-genai==1.29.0
googleapis-common-protos==1.65.0
grpc_interceptor==0.15.4
grpcio==1.68.0
grpcio-tools==1.54.0
h5py==3.12.1
joblib==1.4.2
mock==5.1.0
numpy==1.24.3
packaging==24.1
pandas==2.2.3
pillow==10.4.0
phonenumbers==9.0.10
protobuf==4.25.5
prometheus_client==0.20.0
psutil==6.1.1
PyYAML==6.0.2
requests==2.31.0
ruamel.yaml==0.17.32
rapidfuzz==3.13.0
six==1.16.0
setuptools==75.3.0
scikit-learn==1.5.2
spacy==3.7.5
toml==0.10.2
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.34.0
validators==0.35.0
wordfreq==3.1.1
python-telegram-bot==22.5
tavily-python
ollama